formatted the file structure/home page
was able to get my ideas out and get the file structure sorted for the website. I was also able to get the about/home page to look about the way I want with some minimal updates being needed. not really using anything i read from the svelte docs but hopefully I can do a project using it and have a better understanding of it.